+++ src/ui/general/ask_input_popup.cpp
@@ -1,15 +1,17 @@
 #include "ask_input_popup.hpp"
 
+#include <util/ui.hpp>
+
 using namespace geode::prelude;
 
 bool AskInputPopup::setup(std::string_view title, std::function<void(std::string_view)> function, size_t chars, std::string_view placeholder, std::string_view filter) {
     this->setTitle(std::string(title));
     this->function = function;
 
-    auto winSize = CCDirector::get()->getWinSize();
+    auto rlayout = util::ui::getPopupLayoutAnchored(m_size);
 
     Build<TextInput>::create(m_size.width - POPUP_PADDING * 2.f, std::string(placeholder).c_str(), "chatFont.fnt")
-        .pos(winSize / 2 + CCPoint{0.f, 10.f})
+        .pos(rlayout.fromCenter(0.f, 10.f))
         .parent(m_mainLayer)
         .id("generic-popup-input"_spr)
         .store(input);
@@ -22,7 +24,7 @@ bool AskInputPopup::setup(std::string_view title, std::function<void(std::string
             this->onClose(this);
             this->function(this->input->getString());
         })
-        .pos(winSize / 2 + CCPoint{0.f, -40.f})
+        .pos(rlayout.fromCenter(0.f, -40.f))
         .intoNewParent(CCMenu::create())
         .pos(0.f, 0.f)
         .parent(m_mainLayer);