+++ src/ui/menu/room/room_password_popup.cpp
@@ -4,25 +4,26 @@
 #include <data/packets/client/room.hpp>
 #include <util/format.hpp>
 #include <util/misc.hpp>
+#include <util/ui.hpp>
 
 using namespace geode::prelude;
 
 bool RoomPasswordPopup::setup(uint32_t id) {
     this->setTitle("Join room");
 
-    float popupCenter = CCDirector::get()->getWinSize().width / 2;
+    auto rlayout = util::ui::getPopupLayoutAnchored(m_size);
 
     this->setID("room-password-popup"_spr);
 
     // room password hint
     Build<CCLabelBMFont>::create("Room Password", "bigFont.fnt")
         .scale(0.3f)
-        .pos(popupCenter, POPUP_HEIGHT + 45.f)
+        .pos(rlayout.fromCenter(0.f, 30.f))
         .parent(m_mainLayer);
 
     // room password input node
     Build<TextInput>::create(POPUP_WIDTH * 0.75f, "", "chatFont.fnt")
-        .pos(popupCenter, POPUP_HEIGHT + 25.f)
+        .pos(rlayout.fromCenter(0.f, 10.f))
         .parent(m_mainLayer)
         .store(roomPassInput);
 
@@ -37,7 +38,7 @@ bool RoomPasswordPopup::setup(uint32_t id) {
         .id("join-btn"_spr)
         .intoNewParent(CCMenu::create())
         .id("join-btn-menu"_spr)
-        .pos(popupCenter, 120.f)
+        .pos(rlayout.fromBottom(25.f))
         .parent(m_mainLayer);
 
     return true;